Move coordinator to separate module
